knowshift-backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
supabase>=2.4.0
langchain>=0.2.0
langchain-text-splitters>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.7.0
pdfplumber>=0.11.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic-settings>=2.2.0
httpx>=0.27.0
tenacity>=8.3.0
celery>=5.4.0
redis>=5.0.0
reportlab>=4.1.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.1.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
pytest-cov>=5.0.0
pytest-timeout>=2.3.0
freezegun>=1.4.0
